Refactor _manager.py folder
