+++ taipy/core/_manager/_manager.py
