+++ tests/core/common/test_manager.py
@@ -14,8 +14,8 @@ import pathlib
 from dataclasses import dataclass
 from typing import Any, Dict
 
+from taipy.core._manager._manager import _Manager
 from taipy.core._repository import _FileSystemRepository
-from taipy.core.common._manager import _Manager
 from taipy.core.config.config import Config
